+++ R/explotation.R
@@ -27,6 +27,9 @@ get_level_geometries <- function(gd,
 get_level_geometries.geodimension <- function(gd,
                                               level_name = NULL) {
   stopifnot("Missing level name." = !is.null(level_name))
+  if (gd$snake_case) {
+    level_name <- my_to_snake_case(level_name)
+  }
   level_name <-
     validate_names(names(gd$geolevel), level_name, 'level')
   sort(names(gd$geolevel[[level_name]]$geometry))
@@ -88,6 +91,9 @@ select_levels <- function(gd, level_names = NULL) {
 select_levels.geodimension <- function(gd, level_names = NULL) {
 
   stopifnot("Missing level names." = !is.null(level_names))
+  if (gd$snake_case) {
+    level_names <- my_to_snake_case(level_names)
+  }
   level_names <-
     validate_names(names(gd$geolevel), level_names, 'level')
   delete <- setdiff(names(gd$geolevel), level_names)
@@ -148,6 +154,9 @@ get_level_data.geodimension <- function(gd,
                                         inherited = FALSE,
                                         add_prefix = TRUE) {
   stopifnot("Missing level name." = !is.null(level_name))
+  if (gd$snake_case) {
+    level_name <- my_to_snake_case(level_name)
+  }
   level_name <-
     validate_names(names(gd$geolevel), level_name, 'level')
   data <- gd$geolevel[[level_name]]$data
@@ -172,6 +181,66 @@ get_level_data.geodimension <- function(gd,
 
 
 
+#' Set level data
+#'
+#' Set the data table of a given level.
+#'
+#' We can get the table, filter or transform the data and redefine the level table.
+#'
+#' It is checked that the attributes that have been used in the relationships
+#' remain in the table.
+#'
+#' @param gd A `geodimension` object.
+#' @param level_name A string.
+#' @param data A `tibble` object.
+#'
+#' @return A `geodimension` object.
+#'
+#' @family information output functions
+#'
+#' @examples
+#'
+#' ld <- gd_us |>
+#'   get_level_data(level_name = "county",
+#'                  inherited = TRUE)
+#'
+#' gd_us <- gd_us |>
+#'   set_level_data(level_name = "county",
+#'                  data = ld)
+#'
+#' @export
+set_level_data <- function(gd,
+                           level_name,
+                           data) {
+  UseMethod("set_level_data")
+}
+
+#' @rdname set_level_data
+#' @export
+set_level_data.geodimension <- function(gd,
+                                        level_name = NULL,
+                                        data = NULL) {
+  stopifnot("Missing level name." = !is.null(level_name))
+  if (gd$snake_case) {
+    level_name <- my_to_snake_case(level_name)
+    names(data) <- my_to_snake_case(names(data))
+  }
+  level_name <- validate_names(names(gd$geolevel), level_name, 'level')
+  gd$geolevel[[level_name]]$data <- data
+
+  attributes <- names(data)
+  validate_names(attributes, gd$geolevel[[level_name]]$key, 'key')
+  for (l in names(gd$relation[[level_name]])) {
+    validate_names(attributes, gd$relation[[level_name]][[l]]$lower_fk, 'attribute')
+  }
+  for (l in names(gd$relation)) {
+    validate_names(attributes, gd$relation[[l]][[level_name]]$upper_pk, 'attribute')
+  }
+  gd
+}
+
+
+
 #' @rdname get_level_layer
 #'
 #' @export
@@ -182,6 +251,9 @@ get_level_layer.geodimension <- function(gd,
                                          inherited = FALSE,
                                          add_prefix = TRUE) {
   stopifnot("Missing level name." = !is.null(level_name))
+  if (gd$snake_case) {
+    level_name <- my_to_snake_case(level_name)
+  }
   level_name <-
     validate_names(names(gd$geolevel), level_name, 'level')
   if (is.null(geometry)) {
@@ -213,7 +285,8 @@ get_level_layer.geodimension <- function(gd,
 #' have as a prefix the name of the level.
 #'
 #' Additionally, we indicate the names of the fields where longitude and latitude
-#' will be stored, as well as the crs that is used, if it is not WGS84.
+#' will be stored, as well as the crs that is used, if they are different from the
+#' default values.
 #'
 #' @param gd A `geodimension` object.
 #' @param level_name A string.
@@ -231,15 +304,15 @@ get_level_layer.geodimension <- function(gd,
 #'
 #' ld <- gd_us |>
 #'   get_level_data_geo(level_name = "county",
-#'                  inherited = TRUE)
+#'                      inherited = TRUE)
 #'
 #' @export
 get_level_data_geo <- function(gd,
-                           level_name,
-                           inherited,
-                           add_prefix,
-                           lon_lat,
-                           crs) {
+                               level_name,
+                               inherited,
+                               add_prefix,
+                               lon_lat,
+                               crs) {
   UseMethod("get_level_data_geo")
 }
 
@@ -251,8 +324,28 @@ get_level_data_geo.geodimension <- function(gd,
                                             add_prefix = TRUE,
                                             lon_lat = c("intptlon", "intptlat"),
                                             crs = 4326) {
-  data <- get_level_data_geo(gd, level_name, inherited, add_prefix)
-
-
+  lon_lat <- unique(lon_lat)
+  stopifnot("Two attributes must be indicated: longitude and latitude." = length(lon_lat) == 2)
+  if (gd$snake_case) {
+    level_name <- my_to_snake_case(level_name)
+    lon_lat <- my_to_snake_case(lon_lat)
+  }
+  data <- get_level_data(gd, level_name, inherited, add_prefix)
+  layer <- gd$geolevel[[level_name]]$geometry$point
+  if (is.null(layer)) {
+    gd$geolevel[[level_name]] <-
+      complete_point_geometry(gd$geolevel[[level_name]])
+    layer <- gd$geolevel[[level_name]]$geometry$point
+  }
+  if (!is.null(layer)) {
+    layer <- data |>
+      dplyr::left_join(layer, by = gd$geolevel[[level_name]]$key) |>
+      sf::st_as_sf() |>
+      sf::st_transform(crs)
+    data[, lon_lat] <-
+      matrix(unlist(sf::st_geometry(layer)),
+             ncol = 2,
+             byrow = TRUE)
+  }
   data
 }